backend/requirements.txt
django
djangorestframework
django-cors-headers
Pillow
requests
redis
wechatpayv3
django-unfold